- Rewrite lib to typescript - Add support for db adapters - Add db adapter example with lokijs - Update ESLint configuration - Add example using another db adapter - Typings
