+++ .eslintrc
@@ -1,35 +1,49 @@
 {
+  "root": true,
   "env": {
-    "node": true,
+    "node": true
   },
-  "extends": "airbnb/base",
+  "extends": [
+    "plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended"
+  ],
+  "plugins": [
+    "import",
+    "@typescript-eslint/eslint-plugin"
+  ],
+  "parser": "@typescript-eslint/parser",
   "parserOptions": {
-    "sourceType": "script"
+    "project": "./tsconfig.json"
   },
   "globals": {
     "window": true,
-    "document": true,
+    "document": true
   },
   "rules": {
+    "indent": [
+      "error",
+      "tab"
+    ],
+    "@typescript-eslint/no-explicit-any": 0,
     "max-len": 0,
     "arrow-parens": 0,
     "no-console": 0,
     "no-await-in-loop": 0,
     "object-curly-newline": 0,
-    'no-restricted-syntax': [
-      'error',
+    "no-restricted-syntax": [
+      "error",
       {
-        selector: 'ForInStatement',
-        message: 'for..in loops iterate over the entire prototype chain, which is virtually never what you want. Use Object.{keys,values,entries}, and iterate over the resulting array.',
+        "selector": "ForInStatement",
+        "message": "for..in loops iterate over the entire prototype chain, which is virtually never what you want. Use Object.{keys,values,entries}, and iterate over the resulting array."
       },
       {
-        selector: 'LabeledStatement',
-        message: 'Labels are a form of GOTO; using them makes code confusing and hard to maintain and understand.',
+        "selector": "LabeledStatement",
+        "message": "Labels are a form of GOTO; using them makes code confusing and hard to maintain and understand."
       },
       {
-        selector: 'WithStatement',
-        message: '`with` is disallowed in strict mode because it makes code impossible to predict and optimize.',
-      },
-    ],
+        "selector": "WithStatement",
+        "message": "`with` is disallowed in strict mode because it makes code impossible to predict and optimize."
+      }
+    ]
   }
 }
