+++ .gitignore
@@ -1,3 +1,5 @@
+.idea
+lib
 node_modules
 cookies.json
 followed.json
