+++ package.json
@@ -2,10 +2,12 @@
   "name": "instauto",
   "version": "7.1.2",
   "description": "Instagram automation library written in Node.js",
-  "main": "index.js",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
   "scripts": {
-    "lint": "eslint .",
-    "test": "exit 0"
+    "lint": "node_modules/.bin/eslint --fix src",
+    "test": "exit 0",
+    "build": "rimraf lib && tsc"
   },
   "author": "Mikael Finstad <finstaden@gmail.com>",
   "repository": "mifi/instauto",
@@ -14,14 +16,26 @@
     "node": ">=8"
   },
   "dependencies": {
-    "fs-extra": "^5.0.0",
     "lodash": "^4.17.10",
     "user-agents": "^1.0.559"
   },
+  "peerDependencies": {
+    "puppeteer": ">1.19.0"
+  },
   "devDependencies": {
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb": "^16.1.0",
+    "@types/lodash": "^4.14.176",
+    "@types/lokijs": "^1.5.7",
+    "@types/node": "^16.11.6",
+    "@types/puppeteer": "^5.4.4",
+    "@types/user-agents": "^1.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.2.0",
+    "@typescript-eslint/parser": "^5.2.0",
+    "eslint": "^8.1.0",
+    "eslint-config-airbnb-typescript": "^14.0.1",
     "eslint-plugin-import": "^2.9.0",
-    "puppeteer": "^1.19.0"
+    "lokijs": "^1.5.12",
+    "puppeteer": "^1.19.0",
+    "rimraf": "^3.0.2",
+    "typescript": "^4.4.4"
   }
 }
