+++ src/db_adapters/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './adapter.interface';
