+++ src/index.ts
@@ -1,0 +1,3 @@
+export { Instauto } from './bot';
+export * from "./db_adapters";
+export * from "./interfaces";
