+++ src/interfaces/follower.interface.ts
@@ -1,0 +1,9 @@
+export interface Follower {
+  username: string;
+  time: number;
+}
+
+export interface UnFollower extends Follower {
+  failed?: boolean;
+  noActionTaken?: boolean;
+}
