+++ src/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './follower.interface';
+export * from './likedPhoto.interface';
+export * from './logger.interface';
+export * from './bot-options.interface';
