+++ src/interfaces/likedPhoto.interface.ts
@@ -1,0 +1,5 @@
+export interface LikedPhoto {
+  username: string;
+  href: string;
+  time: number;
+}
